Move mocks to where I can share them across a couple posthog test files
